compatible with N

+++ xposed.cpp
@@ -309,7 +309,7 @@ static bool addPathToEnv(const char* name, const char* path) {
         char newPath[4096];
         int neededLength = snprintf(newPath, sizeof(newPath), "%s:%s", path, oldPath);
         if (neededLength >= (int)sizeof(newPath)) {
-            ALOGE("ERROR: %s would exceed %d characters", name, sizeof(newPath));
+            ALOGE("ERROR: %s would exceed %d characters", name, (int) sizeof(newPath));
             return false;
         }
         setenv(name, newPath, 1);
